# openapi-chatgpt.json
{
    "openapi":  "3.1.0",
    "info":  {
                 "title":  "Market Data Bridge",
                 "description":  "Market data API powered by Yahoo Finance (always available) + IBKR account data (when TWS/Gateway is running). Provides real-time quotes, historical bars, options chains, financials, earnings, news, screeners, and brokerage account info.",
                 "version":  "3.0.0"
             },
    "servers":  [
                    {
                        "url":  "https://denae-unbewildering-sarcastically.ngrok-free.dev"
                    }
                ],
    "components":  {
                       "securitySchemes":  {
                                               "BearerAuth":  {
                                                                  "type":  "http",
                                                                  "scheme":  "bearer"
                                                              },
                                               "ApiKeyAuth":  {
                                                                  "type":  "apiKey",
                                                                  "in":  "header",
                                                                  "name":  "X-API-Key"
                                                              }
                                           }
                   },
    "security":  [
                     {
                         "BearerAuth":  [

                                        ]
                     },
                     {
                         "ApiKeyAuth":  [

                                        ]
                     }
                 ],
    "paths":  {
...
                  "/api/history/{symbol}":  {
                                                "get":  {
                                                            "operationId":  "getHistoricalBars",
                                                            "summary":  "Get historical OHLCV bars for a stock",
                                                            "parameters":  [
                                                                               {
                                                                                   "name":  "symbol",
                                                                                   "in":  "path",
                                                                                   "required":  true,
                                                                                   "schema":  {
                                                                                                  "type":  "string"
                                                                                              },
                                                                                   "description":  "Ticker symbol"
                                                                               },
                                                                               {
                                                                                   "name":  "period",
                                                                                   "in":  "query",
                                                                                   "schema":  {
                                                                                                  "type":  "string",
                                                                                                  "default":  "3mo"
                                                                                              },
                                                                                   "description":  "Time period: \"1d\",\"5d\",\"1mo\",\"3mo\",\"6mo\",\"1y\",\"2y\",\"5y\",\"10y\",\"ytd\",\"max\""
                                                                               },
                                                                               {
                                                                                   "name":  "interval",
                                                                                   "in":  "query",
                                                                                   "schema":  {
                                                                                                  "type":  "string",
                                                                                                  "default":  "1d"
                                                                                              },
                                                                                   "description":  "Bar interval: \"1m\",\"5m\",\"15m\",\"1h\",\"1d\",\"1wk\",\"1mo\""
                                                                               }
                                                                           ],
                                                            "responses":  {
                                                                              "200":  {
                                                                                          "description":  "Historical bars",
                                                                                          "content":  {
                                                                                                          "application/json":  {
                                                                                                                                   "schema":  {
                                                                                                                                                  "type":  "object",
                                                                                                                                                  "properties":  {
                                                                                                                                                                     "symbol":  {
                                                                                                                                                                                    "type":  "string"
                                                                                                                                                                                },
                                                                                                                                                                     "count":  {
                                                                                                                                                                                   "type":  "integer"
                                                                                                                                                                               },
                                                                                                                                                                     "bars":  {
                                                                                                                                                                                  "type":  "array",
                                                                                                                                                                                  "items":  {
                                                                                                                                                                                                "type":  "object",
                                                                                                                                                                                                "properties":  {
                                                                                                                                                                                                                   "time":  {
                                                                                                                                                                                                                                "type":  "string"
                                                                                                                                                                                                                            },
                                                                                                                                                                                                                   "open":  {
                                                                                                                                                                                                                                "type":  "number"
                                                                                                                                                                                                                            },
                                                                                                                                                                                                                   "high":  {
                                                                                                                                                                                                                                "type":  "number"
                                                                                                                                                                                                                            },
                                                                                                                                                                                                                   "low":  {
                                                                                                                                                                                                                               "type":  "number"
                                                                                                                                                                                                                           },
                                                                                                                                                                                                                   "close":  {
                                                                                                                                                                                                                                 "type":  "number"
                                                                                                                                                                                                                             },
                                                                                                                                                                                                                   "volume":  {
                                                                                                                                                                                                                                  "type":  "number"
                                                                                                                                                                                                                              }
                                                                                                                                                                                                               }
                                                                                                                                                                                            }
                                                                                                                                                                              }
                                                                                                                                                                 }
                                                                                                                                              }
                                                                                                                               }
                                                                                                      }
                                                                                      }
                                                                          }
                                                        }
                                            },
...
              }
}